cloud_certification_agent/pyproject.toml
[project]
name = "cloud-certification-agent"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.19.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]


[tool.uv]
package = true
